// src/http/generated/lunajoy-server.schemas.ts
export type GetDailyLogsByPeriod200Item = {
  date: string;
  createdAt: string;
  /** @nullable */
  moodRating: {
  mood: 'VERY_SAD' | 'SAD' | 'NEUTRAL' | 'HAPPY' | 'VERY_HAPPY';
} | null;
  /** @nullable */
  anxietyLevel: {
  anxietyLevelInformed: 'CALM' | 'SLIGHTLY_ANXIOUS' | 'MODERATE' | 'HIGH' | 'OVERWHELMED';
  anxietyLevelCalculated: 'CALM' | 'SLIGHTLY_ANXIOUS' | 'MODERATE' | 'HIGH' | 'OVERWHELMED';
} | null;
  /** @nullable */
  sleepPattern: {
  sleepDurationInHours: number;
  sleepQualityRating: 'VERY_POOR' | 'POOR' | 'AVERAGE' | 'GOOD' | 'EXCELLENT';
  /** @nullable */
  disturbances: string | null;
} | null;
  physicalActivity: ({
  durationInMinutes: number;
  activityType: 'WALKING' | 'RUNNING' | 'CYCLING' | 'SWIMMING' | 'YOGA' | 'OTHER';
  intensity: 'LOW' | 'MODERATE' | 'HIGH';
  satisfactionLevel: 'VERY_DISSATISFIED' | 'DISSATISFIED' | 'NEUTRAL' | 'SATISFIED' | 'VERY_SATISFIED';
})[];
  socialInteraction: ({
  interactionType: 'FAMILY' | 'FRIENDS' | 'COLLEAGUES' | 'STRANGERS' | 'WORK' | 'SHOPPING' | 'OTHER';
  satisfactionLevel: 'VERY_DISSATISFIED' | 'DISSATISFIED' | 'NEUTRAL' | 'SATISFIED' | 'VERY_SATISFIED';
  duration: number;
})[];
  /** @nullable */
  stressLevel: {
  stressLevelType: 'LOW' | 'MODERATE' | 'HIGH' | 'EXTREME';
  duration: number;
  /** @nullable */
  situationDescription: string | null;
  /** @nullable */
  stressSource: 'WORK' | 'FAMILY' | 'HEALTH' | 'FINANCES' | 'RELATIONSHIPS' | 'OTHER' | null;
} | null;
  symptoms: ({
  symptomType: 'SADNESS' | 'FATIGUE' | 'IRRITABILITY' | 'LACK_OF_FOCUS' | 'ANXIETY' | 'INSOMNIA' | 'LOSS_OF_INTEREST' | 'RESTLESSNESS';
  symptomSeverity: 'MILD' | 'MODERATE' | 'SEVERE';
  symptomsDescription: string;
})[];
} | null;